% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_alignment.R
\name{bowtie2}
\alias{bowtie2}
\title{bowtie2}
\usage{
bowtie2(fq1, output.dir, sample.name, bowtie.idx, mc.cores=1, run.cmd=TRUE)
}
\arguments{
\item{fq1}{Path to read1 fastq files}

\item{output.dir}{Output directory}

\item{sample.name}{A character vector for the sample names}

\item{bowtie.idx}{Path to bowtie index files}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}
}
\value{
Aligned SAM files
}
\description{
A wrapper function to run bowtie2.
}
\details{
Bowtie 2 is an ultrafast and memory-efficient tool for aligning sequencing reads to long reference sequences. 
         Bowtie2 is used only for single-end sequencing data.
}
\references{
Fast gapped-read alignment with Bowtie 2
}
\seealso{
\url{http://bowtie-bio.sourceforge.net/bowtie2/index.shtml}
}
