% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_alignment.R
\name{build.star.idx}
\alias{build.star.idx}
\title{STAR}
\usage{
build.star.idx(star.idx.dir=file.path(reference.dir, "STAR.idx"), sample.name, align.dir, ref.fa, ref.gtf, sjdbOverhang=100, star_thread_number=8, fasta.idx=TRUE, SJ.idx=FALSE, run.cmd=TRUE)
}
\arguments{
\item{star.idx.dir}{Directory of STAR index files}

\item{sample.name}{A character vector for the sample names}

\item{ref.fa}{Reference fasta file path}

\item{ref.gtf}{Reference gtf file path (e.g., gencode.gtf)}

\item{sjdbOverhang}{A parameter value for the --sjdbOverhang in STAR. Length of the donor/acceptor sequence on each side of the junctions, ideally=(mate_length-1) (default=100)}

\item{fasta.idx}{Indexing reference fasta file (when first indexing => TRUE)}

\item{SJ.idx}{Indexing splicing junction (when second indexing => TRUE)}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{star_thread_number}{A parameter value for --runThreadN in STAR. A numeric value of the number of threads (default: 8)}
}
\value{
STAR reference and splicing junction indexing
}
\description{
A wrapper function to run STAR (runMode genomeGenerate)
}
\details{
Indexing reference fasta file and splicing junction site from fastq files.
}
\references{
STAR: ultrafast universal RNA-seq aligner
}
\seealso{
\url {https://github.com/alexdobin/STAR}
}
