% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_RNA.quantification.R
\name{cufflinks}
\alias{cufflinks}
\title{cufflinks}
\usage{
cufflinks(fns.bam, sample.name, output.dir, cufflinks_thread_number=4, cufflinks.gtf="G", ref.gtf, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fns.bam}{Path to bam files}

\item{output.dir}{Output directory}

\item{sample.name}{A character vector for the sample names}

\item{cufflinks.thread.number}{A parameter value for -p in Cufflinks. A numeric value of the number of threads (default: 4)}

\item{cufflinks.gtf}{If you set -G, Output will not include novel genes and isoforms that are assembled. (default: -G)}

\item{ref.gtf}{Path to reference gtf file}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
mRNA quantification text files
}
\description{
A wrapper function to run Cufflinks for mRNA quantitation
}
\details{
Cufflinks algorithms for transcript assembly and expression quantification are much more accurate with paired-end reads.
}
\references{
http://cole-trapnell-lab.github.io/cufflinks/papers/
}
\seealso{
\url{http://cole-trapnell-lab.github.io/cufflinks/}
}
