% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_trimming.R
\name{cutadapt}
\alias{cutadapt}
\title{cutadapt}
\usage{
cutadpat(fq1, output.dir, adpat.seq="TGGAATTCTCGGGTGCCAAGG", m=1, mc.cores=1, run.cmd=TRUE)
}
\arguments{
\item{fq1}{Path to fastq files}

\item{output.dir}{Output directory}

\item{adapt.seq}{A parameter value for -b in cutadapt. Adapter sequences user wants to remove}

\item{m}{A parameter value for -m in cutadapt. Discards processed reads that are shorter than m option. Reads that are too short before adapter removal are also discarded.}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\description{
A wrapper function to run Cutadapt
}
\details{
Cutadapt finds and removes adapter sequences, primers, poly-A tails and other types of unwanted sequence.
}
\references{
Cutadapt Removes Adapter Sequences from High-Throughput Sequencing Reads
}
\seealso{
\url{https://cutadapt.readthedocs.io/en/stable/}
}
