% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_makeSet.R
\name{eset2SE}
\alias{eset2SE}
\title{eset2SE}
\usage{
eset2SE(eset = NULL, vset = NULL, cset = NULL, Robject.dir)
}
\arguments{
\item{eset}{RNA abundance ExpressionSet}

\item{vset}{Variant ExpressionSet}

\item{cset}{CNV ExpressionSet}

\item{Robject.dir}{Output directory}
}
\description{
Convert ExpressionSet to SummarizedExperiment
}
\details{
SEQprocess also provides SummarizedExperiment data format for comportable data analysis and management
}
