% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_fastQC.R
\name{fastQC}
\alias{fastQC}
\title{fastQC}
\usage{
fastqc(fq1, fq2, output.dir, run.cmd=TRUE)
}
\arguments{
\item{fq1}{Path to read1 fastq files}

\item{fq2}{Path to read2 fastq files}

\item{output.dir}{Output directory}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
Quality check report for sequence data. (e.g., .html)
}
\description{
A wrapper function to run fastQC
}
\details{
FastQC aims to provide a QC report that detects problems originating from either the sequencer or the starting library material.
}
\references{
FastQC: a quality control tool for high throughput sequence data. Andrews S. (2010).
}
\seealso{
\url{http://www.bioinformatics.babraham.ac.uk/projects/fastqc}
}
