% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_variant.call.recalibration.R
\name{gatk.baserecalibrator}
\alias{gatk.baserecalibrator}
\title{gatk.baserecalibrator}
\usage{
gatk.baserecalibrator(fns.bam, output.dir, sample.name, ref.fa, ref.dbSNP, ref.gold_indels, unsafe=FALSE, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fns.bam}{Path to input BAM files}

\item{output.dir}{Output directory}

\item{sample.name}{A character vector for the sample names}

\item{ref.fa}{Reference fasta file}

\item{ref.dbSNP}{Known SNP sites reference(VCF)}

\item{ref.gold_indels}{Known sites to indel variant call format(VCF)}

\item{unsafe}{A parameter value for -U ALLOW_N_CIGAR_READS in GATK. This parameter must be TRUE in RNA-seq data.}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
GATK report file contained recalibration table by read group, quality scores and all the optional covariates. (e.g., .grp)
}
\description{
A wrapper function to run GATK (BaseRecalibrator)
}
\details{
First pass of the base quality score recalibration. Generates a recalibration table based on various covariates. 
         The default covariates are read group, reported quality score, machine cycle, and nucleotide context. This walker 
         generates tables based on specified covariates via by-locus traversal operating only at sites that are in the known sites VCF.
}
\references{
The Genome Analysis Toolkit: a MapReduce framework for analyzing next-generation DNA sequencing data.
}
\seealso{
\url{https://software.broadinstitute.org/gatk/}
}
