% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_makeSet.R
\name{gatk.depthOFcoverage}
\alias{gatk.depthOFcoverage}
\title{gatk.depthOFcoverage}
\usage{
gatk.depthOFcoverage(vcf.dir, annot.dir, Robject.dir, ref.fa, unsafe,
  minBaseQuality = 1, minMappingQuality = 1, run.cmd = TRUE,
  mc.cores = 1)
}
\arguments{
\item{vcf.dir}{Output of variant call step (directory of vcf files)}

\item{annot.dir}{Output of annotation step (directory of .annovar files)}

\item{ref.fa}{Reference fasta file path}

\item{unsafe}{A parameter value for -U ALLOW_N_CIGAR_READS in GATK. This parameter must be TRUE in RNA-seq data.}

\item{minBaseQuality}{Minimum base quality (default=1)}

\item{minMappingQuality}{Minimum mapping quality (default=1)}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to dedicate. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\description{
Calculate the read depth of the position with single nucleotide variations
}
\details{
When creating a vSet, use read depth to determine whether a mutation exists. GATK DepthOfCoverage uses the interval 
         bed file to calculate the depth of the position.
}
