% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_variant.call.GATK.R
\name{gatk.haplotypecaller}
\alias{gatk.haplotypecaller}
\title{gatk.haplotypecaller}
\usage{
gatk.haplotypecaller(fns.bam, output.dir, sample.name, ref.fa, genotyping_mode="DISCOVERY", output_mode="EMIT_VARIANTS_ONLY", stand_call_conf_number=30, unsafe=FALSE, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fns.bam}{Path to BAM files}

\item{output.dir}{Output directory}

\item{sample.name}{A character vector for the sample names}

\item{ref.fa}{Referance fasta file}

\item{genotyping_mode}{A parameter value for --genotyping_mode in GATK. A character vector to determine the alternate alleles to use for genotyping (default: DISCOVERY)}

\item{output_mode}{A parameter value for --output_mode in GATK. A character vector to produces variant calls (default: EMIT_VARIANTS_ONLY)}

\item{stand_call_conf_number}{A parameter value for -stand_call_conf in GATK. A numeric value of The minimum phred-scaled confidence threshold at which variants should be called (default: 30)}

\item{unsafe}{A parameter value for -U ALLOW_N_CIGAR_READS in GATK. This parameter must be TRUE in RNA-seq data.}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
Variant calling format files (.vcf)
}
\description{
A wrapper function to run GATK (HaplotypeCaller)
}
\details{
HaplotypeCaller is capable of calling SNPs and indels simultaneously via local de-novo assembly of haplotypes in an active 
         region.
}
\references{
The Genome Analysis Toolkit: a MapReduce framework for analyzing next-generation DNA sequencing data.
}
\seealso{
\url{https://software.broadinstitute.org/gatk/}
}
