% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_realignment.R
\name{gatk.realign}
\alias{gatk.realign}
\title{gatk.realign}
\usage{
gatk.realign(fn.rmdu.bam, fn.intervals, output.dir, fn.realign.bam, ref.fa, ref.indels, unsafe=FALSE, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fns.bam}{Path to input BAM files}

\item{fns.interval}{Interval list file created by gatk.targetcreator}

\item{output.dir}{Output directory}

\item{sample.name}{A character vector for the sample names}

\item{ref.fa}{Reference fasta file}

\item{ref.gold_indels}{Known sites to indel variant call format(VCF)}

\item{unsafe}{A parameter value for -U ALLOW_N_CIGAR_READS in GATK. This parameter must be TRUE in RNA-seq data.}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
GATK IndelRealigner returns a Realigned bam file (e.g., realign.bam)
}
\description{
A wrapper function to run GATK (IndelRealigner)
}
\details{
Perform local realignment of reads around indels. IndelRealigner takes a coordinate-sorted and indexed BAM and a target 
         intervals file generated by RealignerTargetCreator. IndelRealigner then performs local realignment on reads coincident with 
         the target intervals using consensus from indels present in the original alignment.
}
\references{
The Genome Analysis Toolkit: a MapReduce framework for analyzing next-generation DNA sequencing data.
}
\seealso{
\url{https://software.broadinstitute.org/gatk/}
}
