% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_realignment.R
\name{gatk.targetcreator}
\alias{gatk.targetcreator}
\title{gatk.targetcreator}
\usage{
gatk.targetcreator(fns.bam, output.dir, sample.name, ref.fa, ref.indels, unsafe=FALSE, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fns.bam}{Path to BAM files}

\item{output.dir}{Output directory}

\item{sample.name}{A character vector for the sample names}

\item{ref.fa}{Reference fasta file}

\item{ref.gold_indels}{Known sites to indel variant call format(VCF)}

\item{unsafe}{A parameter value for -U ALLOW_N_CIGAR_READS in GATK. This parameter must be TRUE in RNA-seq data.}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
Interval file included indel positions (e.g., .intervals)
}
\description{
A wrapper function to run GATK (RealignerTargetCreator)
}
\details{
Define insertion and deletion intervals to target for local realignment
}
\references{
The Genome Analysis Toolkit: a MapReduce framework for analyzing next-generation DNA sequencing data.
}
\seealso{
\url{https://software.broadinstitute.org/gatk/}
}
