% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_variant.call.GATK.R
\name{gatk.variantfilter}
\alias{gatk.variantfilter}
\title{gatk.variantfilter}
\usage{
gatk.variantfilter(fns.vcf, output.dir, sample.name, ref.fa, FS=30.0, QD=2.0, QUAL=50, DP=5, gatk.window=35, cluster=3, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fns.vcf}{Path to VCF files}

\item{output.dir}{Output directory}

\item{ref.fa}{Reference fasta file}

\item{FS}{A parameter value for FS in GATK. FisherStrand. (default=30.0)}

\item{QD}{A parameter value for QD in GATK. Quality by Depth. (default=2.0)}

\item{QUAL}{A parameter value for QUAL in GATK. Low quality. (default=50)}

\item{DP}{A parameter value for DP in GATK. Low depth. (default=5)}

\item{gatk.window}{A parameter value for -window in GATK. The window size (in bases) in which to evaluate clustered SNPs.}

\item{cluster}{A parameter value for -cluster in GATK. The number of SNPs which make up a cluster. Must be at least 2.}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}

\item{sample.name}{A character vector for the sample names}
}
\value{
Filtered VCF file (eg. .f.vcf)
}
\description{
A wrapper function to run GATK (VariantFiltration)
}
\details{
Filter variant calls based on INFO and/or FORMAT annotations. This tool is designed for hard-filtering variant calls based 
         on certain criteria.
}
\references{
The Genome Analysis Toolkit: a MapReduce framework for analyzing next-generation DNA sequencing data.
}
\seealso{
\url{https://software.broadinstitute.org/gatk/}
}
