% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_CNV.R
\name{generate.GC}
\alias{generate.GC}
\title{generate.GC}
\usage{
generate.GC(window=1,000,000, output.dir, ref.fa, run.cmd=TRUE)
}
\arguments{
\item{window}{A parameter value for -w in sequenza. Indicate a window size (in bases), to be used for the binning.
The heterozygous positions and the positions carrying variant calls are not affected by binning.}

\item{output.dir}{Output directory}

\item{ref.fa}{Reference fasta file path}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}
}
\description{
A wrapper function to run sequenza-utils.py in sequenza (GC-windows)
}
\details{
Calculation GC contents from reference fasta file
}
\references{
Sequenza: allele-specific copy number and mutation profiles from tumor sequencing data.
}
\seealso{
{https://cran.r-project.org/web/packages/sequenza/vignettes/sequenza.pdf}
}
