% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_report.R
\name{get.annovar.report}
\alias{get.annovar.report}
\title{get.annovar.report}
\usage{
get.annovar.report(annot.dir)
}
\arguments{
\item{annot.dir}{Path to directory with annovar output files}
}
\value{
list of result summary
}
\description{
Creates a data frame using the annovar output files
}
\details{
Provide data frame of annotation information
}
