% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_function.R
\name{get.fns}
\alias{get.fns}
\title{get.fns}
\usage{
get.fns(input.dir, idx)
}
\arguments{
\item{input.dir}{Path to directory including input files}

\item{idx}{Suffix of input files}
}
\value{
Path to the input files
}
\description{
Gets path of input files
}
