% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_report.R
\name{get.proc.report}
\alias{get.proc.report}
\title{get.proc.report}
\usage{
get.proc.report(proc=c("qc", "trim","cutadapt", "bwa-mem", "bwa-aln", "tophat2","star","bowtie2","rmdu","rmdu_b", "realign", "gatk", "varscan2","mutect","muse","somaticsniper","annot","vep", "cufflinks","htseq","make.eSet"),output.dir)
}
\arguments{
\item{proc}{Process name}

\item{output.dir}{Output directory}
}
\description{
Writes report accoding to process
}
\details{
Creates report information according to processing result
}
