% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_function.R
\name{get.process.names}
\alias{get.process.names}
\title{get.process.names}
\usage{
get.process.names(qc, trim.method, align.method, bwa.method , rm.dup, realign, variant.call.method, annotation.method, rseq.quant.method)
}
\arguments{
\item{qc}{As the quality check progresses, qc is added to the report processes.}

\item{trim.method}{As the trimming progresses, trimming method is added to the report processes.}

\item{align.method}{As the alignment progresses, alignment method is added to the report processes.}

\item{bwa.method}{When alignment is performed with bwa, bwa method is added to the report processes}

\item{rm.dup}{As the removal of duplicates progresses, removal method is added to the report processes.}

\item{realign}{As the re-alignment progresses, re-alignment is added to the report processes.}

\item{variant.call.method}{As the variant calling progresses, variant calling method is added to the report processes.}

\item{annotation.methodd}{As the variant annotation progresses, annotation method is added to the report processes.}

\item{rseq.quant.method}{As the RNA quantification progresses, RNA quantification method is added to the report processes.}
}
\description{
Process names to be used in report
}
