% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_report.R
\name{get.qc.report}
\alias{get.qc.report}
\title{get.qc.report}
\usage{
get.qc.report(qc.dir)
}
\arguments{
\item{qc.dir}{Path to directory with fastQC output files}
}
\value{
data frame of result summary
}
\description{
Creates a data frame using a fastq file
}
\details{
Adds information (ex.Q30) using the data frame combined with the fastqc file and output the result as a data frame
}
