% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_report.R
\name{get.Robject.report}
\alias{get.Robject.report}
\title{get.Robject.report}
\usage{
get.Robject.report(Robject.dir)
}
\arguments{
\item{Robject.dir}{Path to Robject directory}
}
\value{
list of data set summary
}
\description{
Reads rda file
}
\details{
Reads information according to data set.
}
