% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_report.R
\name{get.single_end.metrics.report}
\alias{get.single_end.metrics.report}
\title{get.single.end.metrics.report}
\usage{
get.single_end.metrics.report(sam.dir, collectmetrics.idx=".alignment_summary_metrics$")
}
\arguments{
\item{sam.dir}{Path to directory with sam files}

\item{collectmetric.idx}{Index of file (default=".alignment_summary_metrics$")}
}
\value{
data frame of the result summary
}
\description{
Creates a data frame using picard.collectmultiplemetrics()
}
\details{
Creates a data frame using the txt file, the output of  picard.collectmultiplemetrics(). Used for Single-end data(ex. miRSEQ).
}
