% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_report.R
\name{get.tophat.report}
\alias{get.tophat.report}
\title{get.tophat.report}
\usage{
get.tophat.report(tophat.dir)
}
\arguments{
\item{tophat.dir}{Path to directory with bam files}
}
\value{
data frame of result summary
}
\description{
Creates a data frame using the tophat align summary
}
\details{
Creates a data frame using the txt file, the output of Tophat.
}
