% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_report.R
\name{get.trim.gal.report}
\alias{get.trim.gal.report}
\title{get.trim.gal.report}
\usage{
get.trim.gal.report(trim.dir)
}
\arguments{
\item{trim.dir}{Path to directory with trimmed fastq files}
}
\value{
data frame of result summary
}
\description{
Creates a data frame using the trimming output
}
\details{
Creates a data frame using the txt file, the output of Trim galore.
}
