% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_RNA.quantification.R
\name{gff2gr}
\alias{gff2gr}
\title{gff2gr}
\usage{
gff2gr(mir.gff, output.dir)
}
\arguments{
\item{mir.gff}{Directoy stored at reference gff file}

\item{output.dir}{Output directory}
}
\description{
Converts reference gff file to GRanges form
}
