% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_RNA.quantification.R
\name{gtf2gr}
\alias{gtf2gr}
\title{gtf2gr}
\usage{
gtf2gr(ref.gtf, output.dir)
}
\arguments{
\item{ref.gtf}{Directoy stored at reference gtf file (e.g. gencode.v22.gtf)}

\item{output.dir}{Output directory}
}
\description{
Converts reference gtf file to GRanges form to execute FPKM estimation
}
\details{
To normalize the number of reads of each feature calculated in the previous step to the value of FPKM, convert the reference 
         gtf file to GRanges format.
}
