% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_RNA.quantification.R
\name{htseq.add.info}
\alias{htseq.add.info}
\title{htseq.add.info}
\usage{
htseq.add.info(RNAtype="mRNA", count.fns, output.dir, output.dir, mc.cores=1)
}
\arguments{
\item{RNAtype}{RNAtype (default="mRNA")}

\item{output.dir}{Directory stored at FPKM conunt files}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}

\item{fns.count}{count file paths}
}
\description{
Add information to the htseq output file
}
\details{
Adds information to the output file of htseq. (Gene name, chromosome, start position, end position, gene size, FPKM value)
}
