% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_RNA.quantification.R
\name{htseq_count}
\alias{htseq_count}
\title{htseq_count}
\usage{
htseq_count(RNAtype="mRNA", fns.bam, sample.name, output.dir, Mode="intersection-nonempty", stranded="no", idattr="gene_id", htseq.r="pos", htseq.a=10, ref.gtf, mir.gff, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fns.bam}{Path to input BAM or SAM files}

\item{sample.name}{A character vector for the sample names}

\item{output.dir}{Output directory}

\item{stranded}{A parameter value for -s in htseq-count. Whether the data is from a strand-specific assay (default:no)}

\item{idattr}{A parameter value for -i in htseq-count. GFF attribute to be used as feature ID (default:"gene_id")}

\item{htseq.r}{A parameter value for -r in htseq-count. Sorting order method (default:"pos")}

\item{htseq.a}{A parameter value for -a in htseq-count. Skip all reads with alignment quality lower than the given minimum value (default: 10)}

\item{ref.gtf}{Directoy stored at reference gtf file}

\item{mir.gff}{Directoy stored at micro-RNA reference gff file}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}

\item{MODE}{A parameter value for -m in htseq-count. Mode to handle reads overlapping more than one feature (default:intersection-nonempty)}
}
\value{
Text file included read count information
}
\description{
A wrapper function to run htseq-count for mRNA or miRNA quantitation
}
\details{
Counting reads in features. Given a file with aligned sequencing reads and a list of genomic features, a common task is to 
         count how many reads map to each feature.
}
\references{
HTSeq—a Python framework to work with high-throughput sequencing data
}
\seealso{
\url {https://htseq.readthedocs.io/en/release_0.9.1/}
}
