% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_makeSet.R
\name{make.cset}
\alias{make.cset}
\title{make.cset}
\usage{
make.cset(cnv.dir, Robject.dir)
}
\arguments{
\item{cnv.dir}{sequenza output directory}

\item{Robject.dir}{Ouptut directory}
}
\description{
The copy number variants data are transformed to a file with extension .cSet
}
