% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_makeSet.R
\name{make.eset}
\alias{make.eset}
\title{make.eset}
\usage{
make.eset(RNAquant.dir, Robject.dir, mc.cores=1)
}
\arguments{
\item{RNAquant.dir}{Cufflinks or htseq-count output directory}

\item{Robject.dir}{Output directory}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\description{
For the expression data are transformed to a file with extension .eSet
}
\details{
Quantify mRNA and miRNA and store them in ExpressionSet format for convenient analysis
}
