% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_makeSet.R
\name{make.vset}
\alias{make.vset}
\title{make.vset}
\usage{
make.vset(annot.dir, Robject.dir, mc.cores = 1)
}
\arguments{
\item{annot.dir}{ANNOVAR output directory}

\item{Robject.dir}{Ouptut directory}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}

\item{mut.cnt.cutoff}{Depth filter}
}
\description{
The mutations data are transformed to a file with extension .vSet
}
