% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_variant.call.MuSE.R
\name{MuSE.sump}
\alias{MuSE.sump}
\title{MuSE.sump}
\usage{
MuSE.sump(MuSE.txt, output.dir, ref.dbSNP, ref.gold_indels, data.type="E", run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{MuSE.txt}{Path to MuSE.call output text file}

\item{output.dir}{Output directory}

\item{MuSE.data.type}{E is used for WXS data and G can be used for WGS data}

\item{ref.dbSNP}{Known SNP sites reference}

\item{ref.gold_indels}{Known Indel sites reference}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
vcf files included variant call information
}
\description{
A wrapper function to run MuSE (sump)
}
\details{
The second step, MuSE sump, takes as input the output file from MuSE.call and dbSNP variant call format file. MuSE provide 
         two options for building the sample-specific error model. One is applicable to WES data (option ‘-E’), and the other to WGS 
         data (option -G).
}
\references{
MuSE: accounting for tumor heterogeneity using a sample-specific error model improves sensitivity and specificity in mutation calling from sequencing data
}
\seealso{
\url{http://bioinformatics.mdanderson.org/main/MuSE}
}
