% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_variant.call.MuSE.R
\name{Muse.tabix2.vcf}
\alias{Muse.tabix2.vcf}
\alias{tabix.vcf}
\title{VCFtabix}
\usage{
tabix.vcf(vcf.gz.file, run.cmd = TRUE)
}
\arguments{
\item{vcf.gz.file}{Path to dbSNP, indel reference vcf.gz file}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}
}
\description{
Tabix indexes a TAB-delimited genome position file in.tab.bgz.
}
