% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_report.R
\name{mut.type.somatic}
\alias{mut.type.somatic}
\title{mut.type.somatic}
\usage{
mut.type.somatic(df, ref="Ref",alt="Alt")
}
\arguments{
\item{df}{data frame from get.annovar.report()}

\item{ref}{column name of the reference (default="Ref")}

\item{alt}{column name of the alteration (default="Alt")}
}
\value{
data frame with mutation type
}
\description{
Write mutation type
}
\details{
In the data frame, enter the mutation type using reference and alteration
}
