% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_variant.call.MuTect2.R
\name{mutect2}
\alias{mutect2}
\title{mutect2}
\usage{
run.mutect2(output.dir, ref.fa, tumor.bam, normal.bam, pon.vcf, cosmic.vcf, ref.dbSNP, contamination_fraction_to_filter=0.02, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{output.dir}{Output directory}

\item{ref.fa}{Reference fasta file}

\item{tumor.bam}{Tumor sample bam files}

\item{normal.bam}{Bam files form normal samples obtained from a function gatk.mutect.normal}

\item{pon.vcf}{Panel of normal samples obtained from a function gatk.combinedvariant}

\item{cosmic.vcf}{Known variant sites of cosmic database vcf file}

\item{ref.dbSNP}{Known SNP sites reference vcf}

\item{contamination_fraction_to_filter}{A parameter value for --contamination_fraction_to_filter in GATK MuTect2. Fraction of contamination to aggressively remove (default=0.02)}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
VCF files
}
\description{
A wrapper function to run GATK (MuTect2) Processed through the variant calling as tumor-normal pairs.
}
\details{
MuTect2 is designed to produce somatic variant calls only, and includes some logic to skip variant sites that are very 
         clearly germline based on the evidence present in the Normal sample compared to the Tumor sample.
}
\references{
Sensitive detection of somatic point mutations in impure and heterogeneous cancer samples
}
\seealso{
\url{https://software.broadinstitute.org/gatk/documentation/tooldocs/3.8-0/org_broadinstitute_gatk_tools_walkers_cancer_m2_MuTect2.php}
}
