% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_alignment.R
\name{picard.addrg}
\alias{picard.addrg}
\title{picard.addrg}
\usage{
picard.addrg(fns.bam, output.dir, sample.name, RGLB="lC", RGPL="Illumina", RGPU="runbarcode", RGSM, SORT_ORDER="coordinate", VALIDATION_STRINGENCY="LENIENT", CREATE_INDEX=TRUE, tmp.dir=file.path(output.dir, "tmp"), run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fns.bam}{Path to BAM files}

\item{output.dir}{Output directory}

\item{RGLB}{A parameter value for RGLB in picard. A character value of Read Group library (default="LC")}

\item{RGPL}{A parameter value for RGPL in picard. A character value of Read Group platform (default="Illumina")}

\item{RGPU}{A parameter value for RGPU in picard. A character value of Read Group platform unit (default="runbarcode")}

\item{RGSM}{A parameter value for RGSM in picard.  character value of Read Group sample name (default=sample.name)}

\item{SORT_ORDER}{A parameter value for SO in picard. Sort order, a character value of sorting method (default="coordinate")}

\item{VALIDATION_STRINGENCY}{A parameter value for VALIDATION_STRINGENCY in picard. A character value of validation stringency (default="LENIENT")}

\item{CREATE_INDEX}{A parameter value for CREATE_INDEX in picard. A character value whether to create .bam index files (default="true")}

\item{tmp.dir}{Temporary directory path (default=./tmp)}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}

\item{sample.name}{A character vector for the sample names}
}
\value{
BAM files added read groups
}
\description{
A wrapper function to run Picard (AddOrReplaceReadGroups)
}
\details{
This tool enables the user to replace all read groups in the INPUT file with a single new read group and assign 
         all reads to this read group in the output BAM files.
}
\seealso{
\url{http://broadinstitute.github.io/picard/}
}
