% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_picard.collectmultiplemetrics.R
\name{picard.collectmetrics}
\alias{picard.collectmetrics}
\title{picard.collectmetrics}
\usage{
picard.collectmetrics(fns.bam, out.fns, ref.fa, run.cmd=T, mc.cores=1)
}
\arguments{
\item{fns.bam}{Path to BAM files}

\item{ref.fa}{Reference fasta file path}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}

\item{output.dir}{Output directory}
}
\value{
txt files
}
\description{
Provide read alignment information.
}
\details{
Provides a summary of the alignment process using the bam files
}
\references{
Currently there is no journal reference for picard.
}
\seealso{
\url{http://broadinstitute.github.io/picard/}
}
\author{
Ji-Hye Lee
}
