% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_alignment.R
\name{picard.reorder}
\alias{picard.reorder}
\title{picard.reorder}
\usage{
picard.reorder(fns.bam, output.dir, sample.name, ref.fa, ALLOW_INCOMPLETE_DICT_CONCORDANCE=FALSE, ALLOW_CONTIG_LENGTH_DISCORDANCE=FALSE, CREATE_INDEX=TRUE, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fns.bam}{Path to BAM files}

\item{output.dir}{Output directory}

\item{ref.fa}{Reference fasta file (eg. GRCh38.fa)}

\item{ALLOW_INCOMPLETE_DICT_CONCORDANCE}{A parameter value for ALLOW_INCOMPLETE_DICT_CONCORDANCE in picard. Logical, allow discordant contig (default=FALSE)}

\item{ALLOW_CONTIG_LENGTH_DISCORDANCE}{A parameter value for ALLOW_CONTIG_LENGTH_DISCORDANCE in picard. Logical, allow contig of different length (default=FALSE)}

\item{CREATE_INDEX}{A parameter value for CREATE_INDEX in picard. Logical, whether to create .bam index files (default=TRUE)}

\item{tmp.dir}{Temporary directory (default= ./tmp)}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}

\item{sample.name}{A character vector for the sample names}
}
\value{
Reordered BAM files (e.g., .rg.od.bam)
}
\description{
A wrapper function to run Picard (ReordrSam)
}
\details{
ReorderSam reorders reads in a BAM file to match the contig ordering in a provided reference file, as determined by exact
         name matching of contigs. Reads mapped to contigs but absent in the new reference are dropped. Runs substantially faster if the input 
         is an indexed BAM file.
}
\seealso{
\url{http://broadinstitute.github.io/picard/}
}
