% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_removeDuplicates.R
\name{picard.rmdu}
\alias{picard.rmdu}
\title{picard.rmdu}
\usage{
picard.rmdu(fns.bam, output.dir, sample.name, out.metrics, CREATE_INDEX=TRUE, REMOVE_DUPLICATES=TRUE, VALIDATION_STRINGENCY="LENIENT", tmp.dir=file.path(output.dir, "tmp"), BARCODE_TAG=FALSE, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fns.bam}{Path to BAM files}

\item{output.dir}{Output directory}

\item{sample.name}{A character vector for the sample names}

\item{CREATE_INDEX}{A parameter value for CREATE_INDEX in picard. Logical, whether to create bam index files (default=TRUE)}

\item{REMOVE_DUPLICATES}{A parameter value for REMOVE_DUPLICATES in picard. Logical, whether to remove duplicates (default=TRUE)}

\item{VALIDATION_STRINGENCY}{A parameter value for VALIDATION_STRINGENCY in picard. A character value of validation stringency (default="LENIENT")}

\item{tmp.dir}{Temporary directory (default= ./tmp)}

\item{BARCODE_TAG}{A parameter value for BARCODE_TAG in picard. If barcode sequencing data, set this option TRUE. Duplicated BARCODE is removed.}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.#' @details The MarkDuplicates tool works by comparing sequences in the 5 prime positions of both reads and read-pairs in a 
SAM/BAM file. An BARCODE_TAG option is available to facilitate duplicate marking using molecular barcodes. After 
duplicate reads are collected, the tool differentiates the primary and duplicate reads using an algorithm that 
ranks reads by the sums of their base-quality scores.}
}
\value{
Removing duplicate reads in bam files (e.g., .rmdu.bam)
}
\description{
A wrapper function to run Picard (MarkDuplicates)
}
\seealso{
\url{http://broadinstitute.github.io/picard/}
}
