% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_CNV.R
\name{pileup2seqz}
\alias{pileup2seqz}
\title{generate.seqz}
\usage{
pileup2seqz(gc.fn, normal.pileup.gz, window=1000000, tumor.pileup.gz, output.dir, run.cmd=T, mc.cores=1)
}
\arguments{
\item{normal.pileup.gz}{samtools pileup file of normal sample}

\item{window}{A parameter value for -w in sequenza. Indicate a window size (in bases), to be used for the binning.}

\item{tumor.pileup.gz}{samtools pileup file of tumor sample}

\item{fn.gc}{output file of generate.GC function}
}
\description{
A wrapper function to run sequenza-utils.py in sequenza (pileup2seqz, seqz-binning)
}
\details{
A seqz file contains genotype information, alleles and mutation frequency, and other features. This file is used as input 
         for the R-based part of Sequenza.
}
\references{
Sequenza: allele-specific copy number and mutation profiles from tumor sequencing data.
}
\seealso{
{https://cran.r-project.org/web/packages/sequenza/vignettes/sequenza.pdf}
}
