% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_CNV.R
\name{ploidyNcellularity}
\alias{ploidyNcellularity}
\title{ploidyNcellularity}
\usage{
ploidyNcellularity(cnv.dir)
}
\arguments{
\item{cnv.dir}{Output directory}
}
\description{
Calculate ploidy and cellularity
}
\details{
Calculate ploidy and cellularity for each paired-sample and quantify the copy number
}
\references{
Sequenza: allele-specific copy number and mutation profiles from tumor sequencing data.
}
\seealso{
{https://cran.r-project.org/web/packages/sequenza/vignettes/sequenza.pdf}
}
