% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_function.R
\name{print_message}
\alias{print_message}
\title{print_message}
\usage{
print_message(cmd)
}
\arguments{
\item{cmd}{What users want to show as a message}
}
\value{
message()
}
\description{
Show command line
}
