% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_variant.call.VarScan2.R
\name{processSomatic}
\alias{processSomatic}
\title{processSomatic}
\usage{
processSomatic(fns.vcf, output.dir, min_tumor_freq=0.1, max_normal_freq=0.05, p_value=0.07, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fns.vcf}{varscan output file path}

\item{output.dir}{Output directory}

\item{min_tumor_freq}{A parameter value for --min-tumor-freq in varscan2. Minimum variant allele frequency in tumor (default:0.10)}

\item{max_normal_freq}{A parameter value for --max-normal-freq in varscan2. Maximum variant allele frequency in normal (default:0.05)}

\item{p_value}{A parameter value for --p-value in varscan2. P-value for high-confidence calling (default:0.07)}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\description{
A wrapper function to run VarScan2 (processSomatic)
}
\details{
processSomatic will separate a somatic output file by somatic_status (Germline, Somatic, LOH). Somatic mutations will 
         further be classified as high-confidence (.hc) or low-confidence (.lc).
}
\references{
VarScan 2: somatic mutation and copy number alteration discovery in cancer by exome sequencing.
}
\seealso{
\url {http://varscan.sourceforge.net/}
}
