% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_report.R
\name{qc_aggregate}
\alias{qc_aggregate}
\title{qc_aggregate}
\usage{
qc_aggregate(qc.dir, pattern="fastqc.zip$")
}
\arguments{
\item{qc.dir}{Path to directory with fastqc.zip files}

\item{pattern}{Index of file (default="fastqc.zip$")}
}
\description{
Aggregates the information from the fastq file
}
\details{
Combine the fastqc results into one data.
}
