% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_CNV.R
\name{read.pileup.gz}
\alias{read.pileup.gz}
\title{read.pileup.gz}
\usage{
read.pileup.gz(ref.fa, fns.bam, sample.name, output.dir, mapQ=1, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{ref.fa}{Reference fasta file path}

\item{sample.name}{A character vector for the sample names}

\item{mapQ}{Mapping quality (default=1)}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}

\item{normal.bam}{BAM files of normal sample}

\item{tumor.bam}{BAM files of tumor sample}
}
\description{
A wrapper function to run samtools (mpileup)
}
\details{
Generates VCF, BCF or pileup for one or multiple BAM files. Alignment records are grouped by sample (SM) identifiers in @RG header 
         lines. If the sample identifiers are absent, each input file is regarded as one sample.
}
\seealso{
\url {http://www.htslib.org/doc/samtools.html}
}
