% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_makeSet.R
\name{refGene2gr}
\alias{refGene2gr}
\title{refGene2gr}
\usage{
refGene2gr(refGene.path, cnv.dir)
}
\arguments{
\item{refGene.path}{Path to refGene.txt}

\item{cnv.dir}{Copy number variation directory}
}
\description{
Convert gene reference file to GRanges form
}
