% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEQprocess.R
\name{SEQprocess}
\alias{SEQprocess}
\title{SEQprocess}
\usage{
SEQprocess(fastq.dir = NULL, output.dir = file.path(getwd(), "result",
  "SEQprocess_result"), argList = list(program.name = "SEQprocess"),
  project.name = "SEQprocess", type = c("WGS", "WES", "BarSEQ", "RSEQ",
  "miRSEQ"), pipeline = c("none", "GDC", "GATK", "BarSEQ", "Tuxedo",
  "miRSEQ"), mc.cores = 1, run.cmd = TRUE, report.mode = FALSE,
  config.fn = system.file("data/config.R", package = "SEQprocess"),
  qc = TRUE, trim.method = c("trim.galore", "cutadapt", "none"),
  align.method = c("bwa", "bowtie2", "tophat2", "star", "none"),
  build.transcriptome.idx = FALSE, tophat.thread.number = 4,
  bwa.method = c("mem", "aln"), bwa.thread.number = 4,
  star.thread.number = 8, rm.dup = c("MarkDuplicates", "BARCODE", "none"),
  realign = TRUE, variant.call.method = c("none", "gatk", "varscan2",
  "mutect2", "muse", "somaticsniper"), gatk.thread.number = 4,
  annotation.method = c("annovar", "vep", "none"), ref = "hg38",
  rseq.abundance.method = c("none", "cufflinks", "htseq"),
  cufflinks.gtf = c("G", "g"), cufflinks.thread.number = 4,
  RNAtype = c("mRNA", "miRNA"), CNV = FALSE, make.eSet = FALSE,
  eset2SummarizedExperiment = FALSE, mut.cnt.cutoff = 8,
  qc.dir = file.path(output.dir, "00_qc"), trim.dir = file.path(output.dir,
  "01_trim"), align.dir = file.path(output.dir, "02_align"),
  rmdup.dir = file.path(output.dir, "03_rmdup"),
  realign.dir = file.path(output.dir, "04_realign"),
  vcf.dir = file.path(output.dir, "05_vcf"),
  annot.dir = file.path(output.dir, "06_annot"),
  RNAquant.dir = file.path(output.dir, "07_RNAquant"),
  cnv.dir = file.path(output.dir, "08_cnv"),
  Robject.dir = file.path(output.dir, "09_Robject"))
}
\arguments{
\item{fastq.dir}{If the user starts the process with fastq files, set the directory for the fastq files.}

\item{output.dir}{Output directory}

\item{argList}{The argument list used by the user in the shell}

\item{project.name}{User's project name}

\item{type}{Sequence data type}

\item{pipeline}{One of the six pipelines provided by SEQprocess}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{report.mode}{Whether the process is finished and report is generated}

\item{config.fn}{Congifure file path}

\item{qc}{Whether quality check}

\item{trim.method}{Set trimming method}

\item{align.method}{Set alignment method}

\item{build.transcriptome.idx}{(tophat) A transcriptome index and the associated data files (the original GFF file) can be thus reused for multiple TopHat runs with this option, so these files are only created for the first run with a given set of transcripts. (default=FALSE)}

\item{tophat.thread.number}{(tophat) A numeric value of the number of threads}

\item{bwa.method}{(bwa) Set bwa method}

\item{bwa.thread.number}{(bwa) A numeric value of the number of threads}

\item{star.thread.number}{(STAR) A numeric value of the number of threads}

\item{rm.dup}{Set the remove duplicates method}

\item{realign}{Whether realignment}

\item{variant.call.method}{Set variant call method}

\item{annotation.method}{Set variant annotation method}

\item{ref}{(annovar) Set annovar reference version}

\item{rseq.abundance.method}{Set RNA quantification method}

\item{cufflinks.gtf}{(cufflinks) If you set "-G", Output will not include novel genes and isoforms that are assembled.}

\item{cufflinks.thread.number}{(cufflinks) A numeric value of the number of threads}

\item{RNAtype}{(htseq) Choose mRNA or miRNA.}

\item{CNV}{Whether estimate copy number variation}

\item{make.eSet}{Make ExpressionSet R data(RNA expression, Copy number variation, Mutation)}

\item{eset2SE}{Convert ExpressionSet R data to SummarizedExperiment R data}
}
\description{
Run the NGS data processing pipeline
}
