% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_CNV.R
\name{seqz2seg}
\alias{seqz2seg}
\title{seqz2seg}
\usage{
seqz2seg(cnv.dir, window=1,000,000)
}
\arguments{
\item{cnv.dir}{Output directory}

\item{window}{A parameter value for -w in sequenza. Indicate a window size (in bases), to be used for the binning.}
}
\description{
Segmentation to estimate DNA copy number variation.
}
\details{
Normalization of depth ratio and DNA segmentation
}
\references{
Sequenza: allele-specific copy number and mutation profiles from tumor sequencing data.
}
\seealso{
{https://cran.r-project.org/web/packages/sequenza/vignettes/sequenza.pdf}
}
