% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_variant.call.SomaticSniper.R
\name{somaticsniper}
\alias{somaticsniper}
\title{somaticsniper}
\usage{
somaticsniper(ref.fa, tumor.bam, normal.bam, output.dir, sample.name, mapQual=1, LOH=TRUE, Genotype=TRUE, somaticQual=15, somaticMutation=0.01, Theta=0.85, Hap.number=2, Hap.diff=0.001, out.format="vcf", run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{tumor.bam}{Tumor sample bam files}

\item{normal.bam}{Normal sample bam files}

\item{output.dir}{Output directory}

\item{sample.name}{A character vector for the sample names}

\item{ref.fa}{Reference fasta file}

\item{mapQual}{A parameter value for -q in SomaticSniper Filtering reads with mapping quality less than INT (default:1)}

\item{LOH}{A parameter value for -L in SomaticSniper. Do not report LOH variants as determined by genotypes (logical)}

\item{Genotype}{A parameter value for -G in SomaticSniper. Do not report Gain of Referene variants as determined by genotypes (logical)}

\item{somaticQual}{A parameter value for -Q in SomaticSniper. Filtering somatic SNV output with somatic quality less than INT (default:15)}

\item{somaticMutation}{A parameter value for -s in SomaticSniper. Prior probability of a somatic mutation (default:0.01)}

\item{Theta}{A parameter value for -T in SomaticSniper. Theta in maq consensus calling model (default:0.85)}

\item{Hap.number}{A parameter value for -N in SomaticSniper. Number of haplotypes in the sample (default:2)}

\item{Hap.diff}{A parameter value for -r in SomaticSniper. Prior of a difference between two haplotypes (default:0.001)}

\item{out.format}{A parameter value for -F in SomaticSniper. Select output format (vcf or classic) (default:vcf)}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
VCF files
}
\description{
A wrapper function to run SomaticSniper
}
\details{
The purpose of this program is to identify single nucleotide positions that are different between tumor and normal
         (or in theory, any two bam files). It takes a tumor bam and a normal bam and compares the two to determine the differences.
}
\references{
SomaticSniper: identification of somatic point mutations in whole genome sequencing data.
}
\seealso{
\url{http://gmt.genome.wustl.edu/packages/somatic-sniper/}
}
