% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_alignment.R
\name{STAR}
\alias{STAR}
\title{STAR}
\usage{
STAR(STAR.idx, fq1, fq2, sample.name, output.dir, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{star.idx.dir}{Directory of STAR index}

\item{output.dir}{Output directory}

\item{sample.name}{A character vector for the sample names}

\item{fq1}{path to read1 fastq files}

\item{fq2}{path to read2 fastq files}

\item{outFilterMultimapScoreRange}{A parameter value for --outFilterMultimapScoreRange in STAR. The score range below the maximum score for multimapping alignments (default: 1)}

\item{outFilterMultimapNmax}{A parameter value for --outFilterMultimapNmax in STAR. Read alignments will be output only if the read maps fewer than this value, otherwise no alignments will be output (default: 20)}

\item{outFilterMismatchNmax}{A parameter value for --outFilterMismatchNmax in STAR. Alignment will be output only if it has fewer mismatches than this value (default: 10)}

\item{alignIntronMax}{A parameter value for --alignIntronMax in STAR. Maximum intron length (default: 500,000)}

\item{alignMatesGapMax}{A parameter value for --alignMatesGapMax in STAR. Maximum genomic distance between mates (default: 1,000,000)}

\item{sjdbScore}{A parameter value for --sjdbScore in STAR. Extra alignment score for alignmets that cross database junctions (default: 2)}

\item{alignSJDBoverhangMin}{A parameter value for --alignSJDBoverhangMin in STAR. Minimum overhang for annotated junctions (default: 1)}

\item{outFilterMatchNminOverLread}{A parameter value for --outFilterMatchNminOverLread in STAR. Float: outFilterMatchNmin normalized to read length (sum of mates’ lengths for paired-end reads) (default: 0.33)}

\item{outFilterScoreMinOverLread}{A parameter value for --outFilterScoreMinOverLread in STAR. Float: outFilterScoreMin normalized to read length (sum of mates’ lengths for paired-end reads) (default: 0.33)}

\item{sjdbOverhang}{A parameter value for --sjdbOverhang in STAR. >=0: Length of the donor/acceptor sequence on each side of the junctions, if =0, splice junction database is not used (default: 100)}

\item{SJ.detect}{First align, detection of splicing junction (default=TRUE)}

\item{SJ.align}{Second align, mapping reads to fastq files (default=FALSE)}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}

\item{star_thread_number}{A parameter value for --runThreadN in STAR. A numeric value of the number of threads (default: 8)}
}
\value{
Aligned BAM files
}
\description{
A wrapper function to run STAR.
}
\details{
Spliced Transcripts Alignment to a Reference (STAR), which was designed to specifically address many of the challenges of 
         RNA-seq data mapping, and uses a novel strategy for spliced alignments.
}
\references{
STAR: ultrafast universal RNA-seq aligner
}
\seealso{
\url {https://github.com/alexdobin/STAR}
}
