% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_variant.annotation.R
\name{table.annovar}
\alias{table.annovar}
\title{table.annovar}
\usage{
table.annovar(fn.annovar, output.dir, sample.name, annovar.db, ref="hg38", protocol="refGene,cytoBand,genomicSuperDups,esp6500siv2_all,1000g2015aug_all,exac03,avsnp147,ljb26_all,cosmic70", protocol.type="g,r,r,f,f,f,f,f,f", nastring=".", run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fns.annovar}{Path to annovar files}

\item{output.dir}{Output directory}

\item{sample.name}{A character vector for the sample names}

\item{annovar.db.dir}{Path to directory with ANNOVAR database}

\item{ref}{A parameter value for -buildver in ANNOVAR. Specify the genome build version (default: hg38)}

\item{protocol}{A parameter value for -protocol in ANNOVAR Database names in ANNOVAR (default: "refGene,cytoBand,genomicSuperDups,esp6500siv2_all,1000g2015aug_all,exac03,avsnp147,ljb26_all,cosmic70")}

\item{protocol.type}{A parameter value for -operation in ANNOVAR. Strings separated by commas that specify the types of operation for each protocol (g: genome, r: region, f: filter, default="g,r,r,f,f,f,f,f,f")}

\item{nastring}{A parameter value for -nastring in ANNOVAR. Strings to display when a score is not available (default: ".")}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
CSV files from annovar format
}
\description{
A wrapper function to run table_annovar.pl in ANNOVAR
}
\details{
This function takes an input variant file (such as a VCF file) and generate a tab-delimited output file with many columns,
         each representing one set of annotations. Additionally, if the input is a VCF file, the program also generates a new output VCF file
         with the INFO field filled with annotation information.
}
\references{
ANNOVAR: functional annotation of genetic variants from high-throughput sequencing data
}
\seealso{
\url{http://annovar.openbioinformatics.org/en/latest/user-guide/}
}
