% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_alignment.R
\name{tophat2}
\alias{tophat2}
\title{tophat2}
\usage{
tophat2(fq1, fq2, output.dir, sample.name, ref.gtf, bowtie.idx, tophat_thread_number=4, build.transcriptome.idx=FALSE, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fq1}{Path to read1 fastq files}

\item{fq2}{Path to read2 fastq files}

\item{output.dir}{Output directory}

\item{sample.name}{A character vector for the sample names}

\item{ref.gtf}{Path to reference gtf file}

\item{bowtie.idx}{Path to directory with bowtie indexes and a prefix for the bowtie indexes}

\item{build.transcriptome.idx}{A parameter value for --transcriptome-index in tophat2. A transcriptome index and the associated data 
files (the original GFF file) can be thus reused for multiple TopHat runs with this option, so these 
files are only created for the first run with a given set of transcripts. (default=FALSE)}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}

\item{tophat_thread_number}{A parameter value for -p in tophat2. A numeric value of the number of threads (default: 4)}
}
\value{
Aligned BAM files
}
\description{
A wrapper function to run tophat2.
}
\details{
TopHat is a program that aligns RNA-Seq reads to a genome to identify exon-exon splice junctions. It is built on 
         the ultrafast short read mapping program Bowtie.
}
\references{
TopHat2: accurate alignment of transcriptomes in the presence of insertions, deletions and gene fusions
}
\seealso{
\url{https://ccb.jhu.edu/software/tophat/manual.shtml}
}
