% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_trimming.R
\name{trim.gal}
\alias{trim.gal}
\title{trim.gal}
\usage{
trim.gal(fq1, fq2, trim.quality=30, trim.clip_R1=13, trim.clip_R2=13, output.dir, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fq1}{Path to read1 fastq files}

\item{fq2}{Path to read2 fastq files}

\item{output.dir}{Output directory}

\item{trim.quality}{A parameter value for --quality in trimgalore. A numeric value of phred score cutoff to trim (default=30)}

\item{trim.clip_R1}{A parameter value for --clip_R1 in trimgalore. A numeric value of bp to remove adaptor from the 5-prime end of read1 files (default=13)}

\item{trim.clip_R2}{A parameter value for --clip_R2 in trimgalore. A numeric value of bp to remove adaptor from the 5-prime end of read2 files (default=13)}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
Trimmed fastq files (e.g., .val_1.fastq, .val_1.fastq)
}
\description{
A wrapper function to run Trim Galore
}
\details{
Trims low quality bases, and cleans up adapter sequences for paired-end files.
}
\references{
Krueger F. Trim Galore!
}
\seealso{
\url{https://www.bioinformatics.babraham.ac.uk/projects/trim_galore}
}
