% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_variant.call.VarScan2.R
\name{varscan}
\alias{varscan}
\title{varscan}
\usage{
varscan(fn.pileup, output.dir, sample.name, min_coverage_normal=8, min_coverage_tumor=6, min_var_freq=0.10, min_freq_for_hom=0.75, somatic_p_value=0.05, strand_filter=0, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fn.pileup}{samtools mpileup output file path}

\item{output.dir}{Output directory}

\item{sample.name}{A character vector for the sample names}

\item{min_coverage_normal}{A parameter value for --min-coverage-normal in VarScan2. Minimum coverage in normal to call somatic (default:8)}

\item{min_coverage_tumor}{A parameter value for --min-coverage-tumor in VarScan2. Minimum coverage in tumor to call somatic (default:6)}

\item{min_var_freq}{A parameter value for --min-var-freq in VarScan2. Minimum variant frequency to call a heterozygote (default:0.10)}

\item{min_freq_for_hom}{A parameter value for --min-freq-for-hom in VarScan2. Minimum frequency to call homozygote (default:0.75)}

\item{somatic_p_value}{A parameter value for --somatic-p-value in VarScan2. P-value threshold to call a somatic site (default:0.05)}

\item{strand_filter}{A parameter value for --strand-fiter in VarScan2. If set to 1, removes variants with > 90 percent strand bias(default:0)}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\description{
A wrapper function to run VarScan2
}
\details{
VarScan is a platform-independent mutation caller for targeted, exome, and whole-genome sequencing data.
}
\references{
VarScan 2: somatic mutation and copy number alteration discovery in cancer by exome sequencing.
}
\seealso{
\url {http://varscan.sourceforge.net/}
}
