% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_variant.annotation.R
\name{vcf2annovar}
\alias{vcf2annovar}
\title{vcf2annovar}
\usage{
vcf2annovar(fns.vcf, output.dir, sample.name, format="vcf4", coverage=0, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fns.vcf}{Path to VCF files}

\item{output.dir}{Output directory}

\item{sample.name}{A character vector for the sample names}

\item{format}{A parameter value for -format in ANNOVAR Input files format (.vcf)}

\item{coverage}{A parameter value for -coverage in ANNOVAR Read coverage threshold in pileup file (default:0)}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
Converted annovar files from variant calling format (e.g., .annovar)
}
\description{
A wrapper function to run vcf2annovar.pl in ANNOVAR
}
\details{
The convert2annovar.pl script convert other "genotype calling" format into ANNOVAR format.
         Additionally, the program can generate ANNOVAR input files from a list of dbSNP identifiers, or from transcript identifiers, 
         or from a genomic region.
}
\references{
ANNOVAR: functional annotation of genetic variants from high-throughput sequencing data
}
\seealso{
\url{http://annovar.openbioinformatics.org/en/latest/user-guide/}
}
