% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_variant.annotation.R
\name{vep}
\alias{vep}
\title{Vairant Effect Predictor (VEP)}
\usage{
vep(fns.vcf, output.dir, sample.name, perl5.10.path="/usr/bin", vep.db.dir, run.cmd=TRUE, mc.cores=1)
}
\arguments{
\item{fns.vcf}{Path to VCF files}

\item{output.dir}{Output directory}

\item{vep.db.dir}{Specify the cache directory to use.}

\item{perl5.10.path}{Absolute path to perl 5.10 version. VEP is a Perl based tool. We recommend version 5.10. (default="/usr/bin")}

\item{run.cmd}{Whether to execute the command line (default=TRUE)}

\item{mc.cores}{The number of cores to use. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\value{
text file and html file included variant information
}
\description{
A wrapper function to run VEP
}
\details{
The VEP uses the coordinates and alleles in the VCF file to infer biological context for each variant including the location 
         of each mutation, its biological consequence (frameshift/ silent mutation), and the affected genes. Variants in the VCF files 
         are also matched to known variants from external mutation databases.
}
\references{
The Ensembl Variant Effect Predictor
}
\seealso{
\url{https://asia.ensembl.org/info/docs/tools/vep/script/vep_options.html}
}
