% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_makeSet.R
\name{vset.preprocess}
\alias{vset.preprocess}
\title{vset.preprocess}
\usage{
vset.preprocess(Robject.dir, mut.cnt.cutoff)
}
\arguments{
\item{annot.dir}{Output of annotation step (directory of .annovar files)}

\item{Robject.dir}{Ouptut directory}

\item{mut.cnt.cutoff}{Criterion of depth (default=8)}

\item{mc.cores}{The number of cores to dedicate. Must be at least one(default=1), and parallelization requires at least two cores.}
}
\description{
Use the depth of the variants position to determine the presence or absence of the mutation.
}
\details{
vSet before preprocessing is simply denoted as 1 if there is a mutation and 0 otherwise. After the depth of the mutated position
         is calculated, the variants position with a depth that does not meet a certain criterion is not included in the analysis.
}
